Model second interation and Module first iteration.
The service abstraction is included in the model. Then the needs arise to start wiring the different elements (controllers, services, etc.)
in flexible and maintenable way.

The solution is the module, which is the object responsible for instantiating the different objects and wiring their dependencies and configuration
parameters in a centralized location in the source code.